Add Logout code

+++ core/src/main/java/com/joolove/core/security/oauth2/CommonLogoutSuccessHandler.java
@@ -1,48 +1,46 @@
 package com.joolove.core.security.oauth2;
 
+import com.joolove.core.domain.auth.RefreshToken;
 import com.joolove.core.domain.member.User;
+import com.joolove.core.security.jwt.utils.JwtUtils;
 import com.joolove.core.security.service.LogoutTokenService;
 import com.joolove.core.security.service.RefreshTokenService;
 import com.joolove.core.security.service.UserPrincipal;
+import com.joolove.core.service.UserService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.annotation.Order;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.logout.LogoutHandler;
-import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
+import org.springframework.stereotype.Component;
 
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
+import java.util.Optional;
 
+
+@Order(101)
+@Component
+@RequiredArgsConstructor
 public class CommonLogoutSuccessHandler implements LogoutHandler {
 
-    @Autowired
-    private RefreshTokenService refreshTokenService;
-    @Autowired
-    private LogoutTokenService logoutTokenService;
+    private final RefreshTokenService refreshTokenService;
+    private final LogoutTokenService logoutTokenService;
+    private final UserService userService;
+    private final JwtUtils jwtUtils;
 
     @Override
     public void logout(HttpServletRequest request, HttpServletResponse response, Authentication authentication) {
+        String jwt = jwtUtils.getJwtFromCookies(request);
+        String refreshToken = jwtUtils.getJwtRefreshFromCookies(request);
 
-        // authentication 객체가 null이면 로그인이 안된 상태라고함
-        // 그럼 logout 처리가 되면서 이 handler 보다 빨리 처리가 되어 null 으로 바뀐것으로 예상.
-        // 해결방법은 더 빠른 filter나 handler 에서 아래 과정을 처리하거나 request로 온 header 에서 jwt 쿠키 남아있으면 이걸로 해결하기
+        if (refreshTokenService.findByToken(refreshToken).isPresent()) {
+            RefreshToken token = refreshTokenService.findByToken(refreshToken).get();
+            User user = userService.findByUsername(token.getUser().getUsername());
 
-
-        UserPrincipal userPrincipal = (UserPrincipal) authentication.getPrincipal();
-        User user = userPrincipal.getUser();
-
-        logoutTokenService.createRefreshToken(user.getId());
-        refreshTokenService.deleteByUser(user);
-
-        SecurityContextHolder.getContext().setAuthentication(null);
-        SecurityContextHolder.clearContext();
-
-        try {
-            response.sendRedirect("/");
-        } catch (IOException e) {
-            throw new RuntimeException(e);
+            logoutTokenService.createRefreshToken(user.getId());
+            refreshTokenService.deleteByUser(user);
         }
     }
 }
